xilpm: versal: Optimize XPm_ClockNode data structure
Use u16 for UseCount instead of u32 and use u8 for NumParents instead
of u16 in XPm_ClockNode data structure to reduce the existing buffer
requirement of LibPM.

This reduces around 0.5k bytes of existing buffer requirement.


Acked-for-series: Jolly Shah <jolly.shah@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_clock.c
@@ -275,7 +275,7 @@ done:
 	return Status;
 }
 
-XStatus XPmClock_AddParent(u32 Id, u32 *Parents, u32 NumParents)
+XStatus XPmClock_AddParent(u32 Id, u32 *Parents, u8 NumParents)
 {
 	XStatus Status = XST_SUCCESS;
 	u32 Idx = 0;
